% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayer_linear.R
\name{bayer_linear}
\alias{bayer_linear}
\title{Linear Regression Model Bayesian analysis}
\usage{
bayer_linear(
  data,
  outcome,
  robust = TRUE,
  covariates = NULL,
  random_covariates = NULL,
  sigma_covariates = NULL,
  sigma_random_covariates = NULL,
  sigma_link = "log",
  nu_link = "logm1",
  cores = NULL,
  priors = NULL,
  weights = NULL,
  brm_chains = NULL,
  brm_threads = NULL,
  brm_sig_figs = 10,
  brm_iter = 2000,
  brm_thin = 1,
  brm_warmup = base::floor(brm_iter/2),
  brm_seed = NA,
  brm_adapt_delta = 0.999,
  brm_max_treedepth = 10,
  print = FALSE
)
}
\arguments{
\item{data:}{the data.frame input object}

\item{outcome:}{the column name of the outcome variable (please note: it must be complete, no NA allowed)}

\item{robust:}{if set to FALSE a simple Gaussian family regression will be operated, otherwise a Student family will be used, more robust against ouliers (default)}

\item{covariates:}{vector of model covariate column names and additional terms, such as "Intercept", "0", etc. (optional)}

\item{random_covariates:}{vector of model covariate to use as random effect terms (optional)}

\item{sigma_link:}{link function to be used for sigma analysis (defaults to "log")}

\item{nu_link:}{link function to be used for nu analysis (defaults to "logm1")}

\item{sigma_covariates:}{vector of model covariates to use for sigma specification (optional)}

\item{sigma_rnadom_covariates:}{vector of model covariates to use as random effects for sigma specification (optional)}

\item{cores:}{the number of processor cores to parallelize analysis. It requires cmdstanr installed and configured. brm recommends setting the ‘mc.cores’ option to be as many processors as the hardware and RAM allow (optional)}

\item{priors:}{the vector with prior probability distribution for the Bayesian analysis (optional). These are "brmsprior" objects created by "brms::set_prior()" or related functions and combined using the "c()" method or the "+" operator. See also "brms::default_prior" for more help. brm() prior option}

\item{weights:}{the column name of the weights (e.g. IPTW) to apply to the analysis (optional)}

\item{brm_chains:}{number of Markov chains used in brm() (defaults to 4)}

\item{brm_threads:}{number of threads to use in within-chain parallelization: brm() threads option (optional)}

\item{brm_sig_figs:}{number of decimal digits to store the posterior draws (defaults to 10)}

\item{brm_iter:}{number of total iterations per chain (including warmup; defaults to 2000): brm() iter option}

\item{brm_thin:}{thinning rate. Must be a positive integer. Set ‘thin > 1’ to save memory and computation time if "brm_iter" is large: brm() thin option}

\item{brm_warmup:}{a positive integer specifying number of warmup (aka burnin) iterations. This also specifies the number of iterations used for stepsize adaptation, so warmup draws should not be used for inference. The number of warmup should not be larger than "brm_iter" and the default is "brm_iter / 2": brm() warmup option}

\item{brm_seed:}{the seed for random number generation to make results reproducible. If NA (the default), "Stan" will set the seed randomly: brm() seed option}

\item{brm_adapt_delta:}{brm() control factor to regulate the number of divergent transitions that cause a bias in the obtained posterior draws. This should be generally 0.8 to 0.999: brm() control option (default to 0.999). See stan for more details}

\item{brm_max_treedepth:}{brm() control factor to regulate the number of divergent transitions that cause a bias in the obtained posterior draws. This should be generally >=10: brm() control option (defaults to 10). See stan for more details}

\item{print:}{print intermediate passages resume to stdout (defaults to FALSE)}
}
\value{
A brms::brm model with Bayesian posterior draws
}
\description{
This function compile a Linear model and returns a model with posterior probabilities
based on the outcome specified in the input data.frame. The model that is fitted will be
compiled as a brms::brm() model of student (or gaussian) family with several parameters
adaptable based on user preferences. Priors can be specified with "priors" option
and weighted regression can be performed thorugh "weights" option. To fasten up the
analysis "cmdstanr" package can be used, but it needs to be installed and configured.
}
