% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brms_event_var.R
\name{brms_event_var}
\alias{brms_event_var}
\title{Convert an event data.frame column coded as a Surv() event into a brms cens() event}
\usage{
brms_event_var(data, event_var)
}
\arguments{
\item{data:}{the input data.frame object}

\item{event_var:}{the column name of the event variable (please note: it must be complete, no NA allowed)}
}
\value{
A vector() object which is the brns::brmsformula coded event factor
}
\description{
This function converts a time event column coded as a Surv() event:
\itemize{
\item 0 and 1 (event)
\item 1 and 2 (event)
\item 0 (right censored), 1 (event), 2 (left censored) and 3 (interval censored)
into a brms::brmsformula cens() event:
\item "right" (right censored), "none" (event), "left" (left censored) and "interval" (interval censored)
}
}
