% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brms_outcome_var.R
\name{brms_outcome_var}
\alias{brms_outcome_var}
\title{Convert an outcome variable to brms specification}
\usage{
brms_outcome_var(var_vector, positive = NULL, family = "binomial")
}
\arguments{
\item{var_vector:}{a variable vector to be passed as brms::brm() model outcome}

\item{positive:}{the level of the variable to use as 1 in case of "binomia" family (optional)}

\item{family:}{a brms::brmsfamily() (defaults to "binomia")}
}
\value{
A brms::brm() adapted outcome based on family specified
}
\description{
This function takes an outcome variable vector and reconvert it accordingly to
brms::brm() model specifications
}
