% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_bf.R
\name{compare_bf}
\alias{compare_bf}
\title{Extract bayes factor directly comparing model compiled with "rstan" and faster sampling with non-rstan}
\usage{
compare_bf(
  model,
  model_stanr,
  model_null,
  model_null_stanr,
  silent = TRUE,
  verbose = FALSE,
  log = FALSE
)
}
\arguments{
\item{model:}{a brms::brm() model compiled without "rstan"}

\item{model_null:}{a brms::brm() null model compiled without "rstan"}

\item{model_stanr:}{same "model_null" compiled with "rstan"}

\item{silent:}{brms::bayes_factor() option}

\item{verbose:}{brms::bayes_factor() option}

\item{log:}{brms::bayes_factor() option}
}
\value{
A brms::bayes_factor() object from models comparison
}
\description{
This function extract model compiled with "rstan" and sampling with other mothods
to immediately extract bayes factor
}
