% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_null_model.R
\name{compile_null_model}
\alias{compile_null_model}
\title{Compile a null model taking out a variable of interest in the passed model}
\usage{
compile_null_model(
  model,
  var,
  stanr = FALSE,
  model_stanr = NULL,
  model_null_stanr = NULL,
  stdout = FALSE
)
}
\arguments{
\item{model:}{a brms::brm() fitted model}

\item{var:}{instead of passing model and model_null the variable of interest can be passed. This will take longer as model_null will be compiled (optional)}

\item{stanr:}{if the model was compiled with a different backend option than default "rstan", if set to "TRUE" another model and model_null will be compiled with "rstan". This is important to speed up further comparison as in case of get_bayes_factor() (defaults to FALSE)}

\item{model_stanr:}{user can directly pass the model_null compiled with "rstan" (optional)}

\item{stdout:}{print a summary of comparison to stdout (default to TRUE)}
}
\value{
A list() object composed of: 'stanr': the passed model compiled with rstan (if asked); 'null': model_null brms::brm() fitted object; 'null_rstan': model_null compiled with rstan (if asked)
}
\description{
This function compile and sample a null model which is a brms::brm() identical to the passed "model"
without the variable of interest ("var"), in order to use the null model for further comparison and
identify the actual role of the variable of interest in model accuracy and power.
}
