% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_coeff.R
\name{extract_coeff}
\alias{extract_coeff}
\title{Extract and transform regression coefficients for a variable of interest in the model}
\usage{
extract_coeff(model, var, multinomial_level = NULL)
}
\arguments{
\item{model:}{a brms::brm() fitted model}

\item{var:}{the column name of the variable of interest. It must be a variable passed in the model}

\item{multinomial_level:}{the level of interest in the outcome variable if a multinomial model was fitted (optional)}
}
\value{
A list() object composed of "$coef": the variable coefficient (transformed); "$err": the variable coefficien error (trnasformed); "$low95": the variable coefficient low 95\% CI (transformed); "$high95": the variable coefficient high 95\% CI (transformed)
}
\description{
This function extract and transform regression coefficients for a variable of interest
within a brms::fit() model
}
