% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_posterior_stats.R
\name{extract_posterior_stats}
\alias{extract_posterior_stats}
\title{Extract posterior summary statistics given the vector of posterior draws of a specific variable inside a brms::brm() fitted model}
\usage{
extract_posterior_stats(post)
}
\arguments{
\item{post:}{a vector of posterior draws}
}
\value{
A data.frame() object with the summary statistics for the varibale
}
\description{
This function extract summary statistics given the vector of posterior draws of a
variable inside a fitted brms::brm() model
}
