% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_covariates.R
\name{get_covariates}
\alias{get_covariates}
\title{Extract covariates from a regression formula}
\usage{
get_covariates(formula)
}
\arguments{
\item{formula:}{string of the regression formula}
}
\value{
A vector with covatiates of the formula
}
\description{
This function extract covriates from a regression formula: y ~ x1 + x2 + x3
}
