% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo_analysis.R
\name{loo_analysis}
\alias{loo_analysis}
\title{Calculates Leave-One-Out Cross-Validation (LOO) from a brms::brm() fitted model for a variable of interest in the model}
\usage{
loo_analysis(model, model_null = NULL, var = NULL, stdout = TRUE)
}
\arguments{
\item{model:}{a brms::brm() fitted model}

\item{model_null:}{a brms::brm() fitted model without the variable of interest (optional)}

\item{var:}{instead of passing model and model_null the variable of interest name can be passed. This will take longer as model_null will be compiled (optional)}

\item{stdout:}{print a summary of comparison to stdout (default to TRUE)}
}
\value{
A list() object composed of "mean": the mean LOO difference value between model and model_nulll (if negative "model" fits better); "se": the standard error of the LOO value between model and model_nulll; "p": the p-value of LOO comparison between model and model_nulll
}
\description{
This function calculates LOO comparing a brms::brm() fitted model ("model") with the variable
of interest and a null model identical to the passed one without the variable of interest.
This function will also calculate Pareto K values to hel identify obsrvations with issues.
User can either pass the variable of interest name ("var") OR the null model directly ("model_null").
In this case, analysis will take longer as the null model needs to be compiled and sampled.
}
\details{
Interpreting LOO results: When interpreting the results, a significant difference in LOO
(considering the uncertainty in the estimates) implies that one model provides a better fit
to the data than the other.
If the model with the variable of interest has a significantly lower LOO, it suggests that
including that variable improves the model's predictive accuracy. However, if there's no
clear difference, it may indicate that this specific variable doesn't contribute significantly
to the model in terms of predictive power.
If returned "mean" is negative it means "model" fit better.

Interpret the Pareto k Values: The Pareto k values help you identify observations with potential issues.
The values are generally interpreted as follows:
\itemize{
\item k < 0.5: The observation does not pose any issue for LOO.
\item 0.5 ≤ k < 0.7: The observation might be slightly influential; interpret LOO results with caution.
\item 0.7 ≤ k < 1: The observation is highly influential and might affect the reliability of the LOO estimate.
\item k ≥ 1: The observation is very influential and the LOO results are likely unreliable for this observation.
}
}
