% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_imputation.R
\name{missing_imputation}
\alias{missing_imputation}
\title{Operate missing imputation on a data.frame object}
\usage{
missing_imputation(
  data,
  id = NULL,
  table = NULL,
  table_sep = "\\t",
  cores = NULL,
  imp_num = 5,
  seed = NA,
  maxit = 50,
  parallelseed = 500
)
}
\arguments{
\item{data:}{the input data.frame to perform analysis on}

\item{id:}{column name of the ID column. This is important to indicate to the function as imputing on ID column will significantly slow down the imputation. (optional)}

\item{table:}{the full path of the file to write down the iumputed table (optional)}

\item{table_sep:}{the field separator to use for the file to write down the iumputed table (defaults to '\t')}

\item{cores:}{the number of processor cores to parallelize analysis. If null the function will use all available cores (optional)}

\item{imp_num:}{The number of desired imputated datasets, "m" mice::futuremice() option (defaults to 5)}

\item{seed:}{A scalar to be used as the seed value for the mice algorithm, "seed" mice::futuremice() option (optional)}

\item{maxit:}{A scalar giving the maximum number of iterations, "maxit" mice::futuremice() option (defaults to 500)}

\item{parallelseed:}{A scalar to be used to obtain reproducible results, "parallelseed" mice::futuremice() option (optional)}
}
\value{
A data.frame object with imputed missing values. The ID column (if passed) will be the first column in the returned data.frame
}
\description{
This function operates a missing imputation using mice::futuremice and will return
an imputed data.frame.
It is important to specify "id" column name as this will significantly slow down the analysi
if not passed. "table" allows to specify a path to write down the imputed data.frame
}
