% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recompile_stanr.R
\name{recompile_stanr}
\alias{recompile_stanr}
\title{Convert a model compiled without "rstan" to "rstan"}
\usage{
recompile_stanr(model)
}
\arguments{
\item{model:}{a brms::brm() model}
}
\value{
A brms::brm() model compiled with the default compiler "stanr"
}
\description{
This function recompile a model compiled without "stanr" default compiler
with "stanr", in order to speed up get_bayes_factor()
}
