% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/related_factor_variable.R
\name{related_factor_variable}
\alias{related_factor_variable}
\title{Creates a factor variable which is idenical to another factor variable changing only a specified proportion}
\usage{
related_factor_variable(var_vector, prop = 10, stdout = TRUE)
}
\arguments{
\item{var_vector:}{the input factor variable}

\item{prop:}{the specified proportion of the variable to change (defaults to 10)}

\item{stdout:}{print a summary of comparison to stdout (default to TRUE)}
}
\value{
A vector() factor object with the specified proportion of the factor variable changed
}
\description{
This function is primarily used to testing and creates a factor variable which is associated with
another factor variable passed as input, changing only a specified proportion ("propr" 0-100) of
the values.
}
