% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waic_analysis.R
\name{waic_analysis}
\alias{waic_analysis}
\title{Calculates Watanabe-Akaike Information Criterion (WAIC) from a brms::brm() fitted model for a variable of interest in the model}
\usage{
waic_analysis(model, model_null = NULL, var = NULL, stdout = TRUE)
}
\arguments{
\item{model:}{a brms::brm() fitted model}

\item{model_null:}{a brms::brm() fitted model without the variable of interest (optional)}

\item{var:}{instead of passing model and model_null the variable of interest name can be passed. This will take longer as model_null will be compiled (optional)}

\item{stdout:}{print a summary of comparison to stdout (default to TRUE)}
}
\value{
A list() object composed of "mean": the mean WAIC difference value between model and model_nulll; "se": the standard error of the WAIC value between model and model_nulll; "p": the p-value of WAIC comparison between model and model_nulll
}
\description{
This function calculates WAIC comparing a brms::brm() fitted model ("model") with the variable
of interest and a null model identical to the passed one without the variable of interest.
User can either pass the variable of interest name ("var") OR the null model directly ("model_null").
In this case, analysis will take longer as the null model needs to be compiled and sampled.
}
\details{
Interpreting WAIC results: When interpreting the results, a significant difference in WAIC
(considering the uncertainty in the estimates) implies that one model provides a better fit
to the data than the other.
If the model with the variable of interest has a significantly lower WAIC, it suggests that
including that variable improves the model's predictive accuracy. However, if there's no
clear difference, it may indicate that this specific variable doesn't contribute significantly
to the model in terms of predictive power.
}
